Address review feedback on #32
- Extract shared binary serialization (vectorDataToBinary, binaryToVectorData)
  and JSON serialization (vectorDataToSerializable, serializableToVectorData)
  into serialization.ts; remove duplicates from file-system-adapter,
  opfs-adapter, and chrome-storage-adapter
- Remove duplicate SerializedVectorData interface from chrome-storage-adapter
  and opfs-adapter, reuse the shared export from serialization.ts
- Fix IndexedDB adapter getMany to catch all BatchOperationError instances
  (not just succeeded === 0) for consistent partial-miss behavior
- Align CI Bun version (1.13.0) with package.json engines constraint
- Fix SQLite adapter exists() to handle both null and undefined from .get()
- Fix browser E2E build naming to emit flat dist/storage.js
- Restore Bun.RedisClient properly in test cleanup when original was undefined
- Remove dead errors array from OPFS adapter getMany
