+++ .github/workflows/verify.yml
@@ -28,7 +28,7 @@ jobs:
       - name: Setup Bun
         uses: oven-sh/setup-bun@v1
         with:
-          bun-version: '1.3.2'
+          bun-version: '1.13.0'
 
       - name: Cache Bun dependencies and install cache
         uses: actions/cache@v4
