+++ src/storage/adapters/opfs-adapter.ts
@@ -5,7 +5,14 @@ import type {
   StorageAdapter,
   VectorData,
 } from '@/core/types.js';
-import { calculateMagnitude } from './serialization.js';
+import {
+  type SerializedVectorData,
+  binaryToVectorData,
+  calculateMagnitude,
+  serializableToVectorData,
+  vectorDataToBinary,
+  vectorDataToSerializable,
+} from './serialization.js';
 
 // OPFS types declared inline since they may not be in the TypeScript lib.
 
@@ -42,207 +49,6 @@ interface FileSystemHandle {
 interface OPFSStorageAdapterOptions {
   directory: string;
   format?: 'binary' | 'json';
-}
-
-// Serialization helpers
-
-interface SerializedVectorData {
-  id: string;
-  vector: number[];
-  metadata?: Record<string, unknown>;
-  magnitude: number;
-  format?: string;
-  normalized?: boolean;
-  timestamp: number;
-  lastAccessed?: number;
-  accessCount?: number;
-  compression?: VectorData['compression'];
-}
-
-function vectorDataToSerializable(data: VectorData): SerializedVectorData {
-  const serialized: SerializedVectorData = {
-    id: data.id,
-    vector: Array.from(data.vector),
-    magnitude: data.magnitude,
-    timestamp: data.timestamp,
-  };
-
-  if (data.metadata !== undefined) {
-    serialized.metadata = data.metadata;
-  }
-  if (data.format !== undefined) {
-    serialized.format = data.format;
-  }
-  if (data.normalized !== undefined) {
-    serialized.normalized = data.normalized;
-  }
-  if (data.lastAccessed !== undefined) {
-    serialized.lastAccessed = data.lastAccessed;
-  }
-  if (data.accessCount !== undefined) {
-    serialized.accessCount = data.accessCount;
-  }
-  if (data.compression !== undefined) {
-    serialized.compression = data.compression;
-  }
-
-  return serialized;
-}
-
-function serializableToVectorData(serialized: SerializedVectorData): VectorData {
-  const data: VectorData = {
-    id: serialized.id,
-    vector: new Float32Array(serialized.vector),
-    magnitude: serialized.magnitude,
-    timestamp: serialized.timestamp,
-  };
-
-  if (serialized.metadata !== undefined) {
-    data.metadata = serialized.metadata;
-  }
-  if (serialized.format !== undefined) {
-    data.format = serialized.format;
-  }
-  if (serialized.normalized !== undefined) {
-    data.normalized = serialized.normalized;
-  }
-  if (serialized.lastAccessed !== undefined) {
-    data.lastAccessed = serialized.lastAccessed;
-  }
-  if (serialized.accessCount !== undefined) {
-    data.accessCount = serialized.accessCount;
-  }
-  if (serialized.compression !== undefined) {
-    data.compression = serialized.compression;
-  }
-
-  return data;
-}
-
-// Binary format helpers
-// Layout: [4-byte uint32 vector length][Float32Array bytes][UTF-8 JSON for remaining fields]
-
-/**
- * The fields stored in the JSON portion of the binary format (everything except the vector).
- */
-interface BinaryRemainingFields {
-  id: string;
-  magnitude: number;
-  timestamp: number;
-  metadata?: Record<string, unknown>;
-  format?: string;
-  normalized?: boolean;
-  lastAccessed?: number;
-  accessCount?: number;
-  compression?: VectorData['compression'];
-}
-
-function buildBinaryRemainingFields(data: VectorData): BinaryRemainingFields {
-  const fields: BinaryRemainingFields = {
-    id: data.id,
-    magnitude: data.magnitude,
-    timestamp: data.timestamp,
-  };
-
-  if (data.metadata !== undefined) {
-    fields.metadata = data.metadata;
-  }
-  if (data.format !== undefined) {
-    fields.format = data.format;
-  }
-  if (data.normalized !== undefined) {
-    fields.normalized = data.normalized;
-  }
-  if (data.lastAccessed !== undefined) {
-    fields.lastAccessed = data.lastAccessed;
-  }
-  if (data.accessCount !== undefined) {
-    fields.accessCount = data.accessCount;
-  }
-  if (data.compression !== undefined) {
-    fields.compression = data.compression;
-  }
-
-  return fields;
-}
-
-function vectorDataToBinary(data: VectorData): ArrayBuffer {
-  const remaining = buildBinaryRemainingFields(data);
-
-  const jsonBytes = new TextEncoder().encode(JSON.stringify(remaining));
-  const vectorLength = data.vector.length;
-  const vectorBytes = vectorLength * 4; // Float32 = 4 bytes each
-
-  // 4 bytes for length prefix + vector bytes + JSON bytes
-  const totalSize = 4 + vectorBytes + jsonBytes.byteLength;
-  const buffer = new ArrayBuffer(totalSize);
-  const view = new DataView(buffer);
-
-  // Write vector length as uint32
-  view.setUint32(0, vectorLength, true);
-
-  // Write Float32Array bytes
-  const float32View = new Float32Array(buffer, 4, vectorLength);
-  float32View.set(data.vector);
-
-  // Write JSON bytes after the vector
-  const jsonOffset = 4 + vectorBytes;
-  const uint8View = new Uint8Array(buffer, jsonOffset);
-  uint8View.set(jsonBytes);
-
-  return buffer;
-}
-
-function binaryFieldsToVectorData(
-  remaining: BinaryRemainingFields,
-  vector: Float32Array,
-): VectorData {
-  const data: VectorData = {
-    id: remaining.id,
-    vector,
-    magnitude: remaining.magnitude,
-    timestamp: remaining.timestamp,
-  };
-
-  if (remaining.metadata !== undefined) {
-    data.metadata = remaining.metadata;
-  }
-  if (remaining.format !== undefined) {
-    data.format = remaining.format;
-  }
-  if (remaining.normalized !== undefined) {
-    data.normalized = remaining.normalized;
-  }
-  if (remaining.lastAccessed !== undefined) {
-    data.lastAccessed = remaining.lastAccessed;
-  }
-  if (remaining.accessCount !== undefined) {
-    data.accessCount = remaining.accessCount;
-  }
-  if (remaining.compression !== undefined) {
-    data.compression = remaining.compression;
-  }
-
-  return data;
-}
-
-function binaryToVectorData(buffer: ArrayBuffer): VectorData {
-  const view = new DataView(buffer);
-
-  // Read vector length
-  const vectorLength = view.getUint32(0, true);
-  const vectorBytes = vectorLength * 4;
-
-  // Read Float32Array
-  const vector = new Float32Array(buffer.slice(4, 4 + vectorBytes));
-
-  // Read JSON for remaining fields
-  const jsonOffset = 4 + vectorBytes;
-  const jsonBytes = new Uint8Array(buffer, jsonOffset);
-  const jsonString = new TextDecoder().decode(jsonBytes);
-  const remaining = JSON.parse(jsonString) as BinaryRemainingFields;
-
-  return binaryFieldsToVectorData(remaining, vector);
 }
 
 // ID sanitization: percent-encode filesystem-unsafe characters.
@@ -373,7 +179,6 @@ export class OPFSStorageAdapter implements StorageAdapter {
   async getMany(ids: string[]): Promise<VectorData[]> {
     const directory = this.requireVectorsHandle();
     const results: VectorData[] = [];
-    const errors: Array<{ id: string; error: Error }> = [];
     const now = Date.now();
 
     for (const id of ids) {
@@ -388,19 +193,14 @@ export class OPFSStorageAdapter implements StorageAdapter {
 
         results.push(data);
       } catch (error: unknown) {
-        if (isNotFoundError(error)) {
-          errors.push({ id, error: new VectorNotFoundError(id) });
-        } else {
-          errors.push({
-            id,
-            error: error instanceof Error ? error : new Error(String(error)),
-          });
+        // Silently skip missing entries — the StorageAdapter contract returns
+        // the found subset rather than throwing for partial misses.
+        if (!isNotFoundError(error)) {
+          throw error;
         }
       }
     }
 
-    // Return whatever was found (possibly empty). Other adapters return []
-    // when no requested IDs exist, so we follow the same convention.
     return results;
   }
